Introduce read and write queries
