+++ source/Nevermore/IReadTransaction.cs
@@ -1,0 +1,6 @@
+using System;
+
+namespace Nevermore
+{
+    public interface IReadTransaction : IReadQueryExecutor, IDisposable { }
+}
