+++ source/Nevermore/IRelationalStore.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Data;
+using System.Runtime.CompilerServices;
 using System.Text;
 using Nevermore.Mapping;
 
@@ -9,10 +10,17 @@ namespace Nevermore
     {
         string ConnectionString { get; }
         int MaxPoolSize { get; }
-        IRelationalTransaction BeginTransaction(RetriableOperation retriableOperation = RetriableOperation.Delete | RetriableOperation.Select, string name = null);
-        IRelationalTransaction BeginTransaction(IsolationLevel isolationLevel, RetriableOperation retriableOperation = RetriableOperation.Delete | RetriableOperation.Select, string name = null);
         void WriteCurrentTransactions(StringBuilder sb);
         DocumentMap GetMappingFor<T>();
         DocumentMap GetMappingFor(Type type);
+        
+        IReadTransaction BeginReadTransaction(RetriableOperation retriableOperation = RetriableOperation.Delete | RetriableOperation.Select, [CallerMemberName]string name = null);
+        IReadTransaction BeginReadTransaction(IsolationLevel isolationLevel, RetriableOperation retriableOperation = RetriableOperation.Delete | RetriableOperation.Select, [CallerMemberName]string name = null);
+        
+        IWriteTransaction BeginWriteTransaction(RetriableOperation retriableOperation = RetriableOperation.Delete | RetriableOperation.Select, [CallerMemberName]string name = null);
+        IWriteTransaction BeginWriteTransaction(IsolationLevel isolationLevel, RetriableOperation retriableOperation = RetriableOperation.Delete | RetriableOperation.Select, [CallerMemberName]string name = null);
+
+        IRelationalTransaction BeginTransaction(RetriableOperation retriableOperation = RetriableOperation.Delete | RetriableOperation.Select, [CallerMemberName]string name = null);
+        IRelationalTransaction BeginTransaction(IsolationLevel isolationLevel, RetriableOperation retriableOperation = RetriableOperation.Delete | RetriableOperation.Select, [CallerMemberName]string name = null);
     }
 }
