+++ source/Nevermore/IWriteTransaction.cs
@@ -1,0 +1,7 @@
+namespace Nevermore
+{
+    public interface IWriteTransaction : IReadTransaction, IWriteQueryExecutor
+    {
+        void Commit();
+    }
+}
