+++ source/Nevermore/QueryExecutorExtensions.cs
@@ -9,7 +9,7 @@ namespace Nevermore
         /// </summary>
         /// <typeparam name="TDocument">The type of document being queried. Results from the database will be mapped to this type.</typeparam>
         /// <returns>A stream of resulting documents.</returns>
-        public static IQueryBuilder<TDocument> Query<TDocument>(this IQueryExecutor queryExecutor) where TDocument : class, IId
+        public static IQueryBuilder<TDocument> Query<TDocument>(this IReadQueryExecutor queryExecutor) where TDocument : class, IId
         {
             return queryExecutor.TableQuery<TDocument>()
                 // AsType creates an instance `QueryBuilder` without actually modifying the query itself.
