+++ source/Nevermore/RelationalTransactionRegistry.cs
@@ -16,7 +16,7 @@ namespace Nevermore
         // Getting a typed ILog causes JIT compilation - we should only do this once
         static readonly ILog Log = LogProvider.For<RelationalTransactionRegistry>();
 
-        readonly List<RelationalTransaction> transactions = new List<RelationalTransaction>();
+        readonly List<ReadRelationalTransaction> transactions = new List<ReadRelationalTransaction>();
         bool highNumberAlreadyLoggedAtError;
 
         public RelationalTransactionRegistry(SqlConnectionStringBuilder connectionString)
@@ -28,7 +28,7 @@ namespace Nevermore
         public string ConnectionString { get; }
         public int MaxPoolSize { get; }
 
-        public void Add(RelationalTransaction trn)
+        public void Add(ReadRelationalTransaction trn)
         {
             lock (transactions)
             {
@@ -42,7 +42,7 @@ namespace Nevermore
             }
         }
 
-        public void Remove(RelationalTransaction trn)
+        public void Remove(ReadRelationalTransaction trn)
         {
             lock (transactions)
                 transactions.Remove(trn);
@@ -72,7 +72,7 @@ namespace Nevermore
 
         public void WriteCurrentTransactions(StringBuilder sb)
         {
-            RelationalTransaction[] copy;
+            ReadRelationalTransaction[] copy;
             lock (transactions)
                 copy = transactions.ToArray();
